+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.11.0
+- Add today learned module.
+- Add daily learning goal.
+- Fix translation.
+
 ## v0.10.5
 - Add translation of new update panel.
 - Add Change log.